+++ db/migrate/20170616114757_add_external_ref_to_form_field.rb
@@ -1,0 +1,6 @@
+class AddExternalRefToFormField < ActiveRecord::Migration[5.0]
+  def change
+    add_column :form_fields, :external_reference, :string, after: 'email'
+    add_index :form_fields, :external_reference
+  end
+end