+++ spec/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20170615184855) do
+ActiveRecord::Schema.define(version: 20170616114757) do
 
   create_table "addresses", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
     t.integer  "addressable_id"
@@ -217,14 +217,16 @@ ActiveRecord::Schema.define(version: 20170615184855) do
 
   create_table "form_fields", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
     t.integer  "form_id"
-    t.text     "locales",    limit: 65535
+    t.text     "locales",            limit: 65535
     t.string   "identifier"
     t.string   "field_type"
     t.boolean  "presence"
     t.boolean  "email"
+    t.string   "external_reference"
     t.integer  "position"
-    t.datetime "created_at",               null: false
-    t.datetime "updated_at",               null: false
+    t.datetime "created_at",                       null: false
+    t.datetime "updated_at",                       null: false
+    t.index ["external_reference"], name: "index_form_fields_on_external_reference", using: :btree
     t.index ["form_id"], name: "index_form_fields_on_form_id", using: :btree
   end